This push implements the following: - Profile Page (personal.js) - Modals for PersonalData (PersonalDataModal.js), FilteringFood (FilterFoodModal.js), Theming (PersonalizeModal.js) in seperate Files - ReactHooks for Managing Dashboard and Foodfilter (food-filter.js) - Cleaned up AppTabbar slightly (AppTabbar.js)
