+++ rogue-thi-app/styles/FilterFoodModal.module.css
@@ -1,0 +1,11 @@
+.filterHeader h6 {
+    float: left;
+}
+
+.filterHeader button {
+    float: right;
+}
+
+.filterBody {
+    clear: both;
+}
